requirements.txt
django
djangorestframework
drf-yasg
gunicorn
polls
markdown
django-filter
psycopg2
django-environ
django-cors-headers
django-rest-auth 
django-allauth
djangorestframework-jwt
pyrebase4
firebase-admin
plotly==5.9.0
kaleido
numpy
pandas
nltk
BeautifulSoup4
lxml
matplotlib
wordcloud
scikit-learn
transformers==4.16.2
tensorflow
spacy
facebook-sdk